// payload-types.ts
export interface Product {
  id: string;
  name: string;
  price: string;
  batchNumber: number;
  country: 'Korea' | 'China' | 'Japan';
  status: 'ordered_to_seller' | 'in_progress' | 'arrived_wh_ln' | 'otw_ina' | 'arrived_wh_ina' | 'arrived_admin';
  isPaid?: boolean | null;
  notes?: string | null;
  updatedAt: string;
  createdAt: string;
}